Move the create_worker call from the init/1 callback
By attempting to connect within the init/1 callback,
we have no room to reconnect, as the init raises, causing
the KafkaEx.ConsumerGroup.init/1 to fail when attempting
to supervise a worker, this then crashes the application.

Save the worker_opts for a new GenServer callback to use

No need to set member_id to nil within init/1
as it will already be nil

Handle create_worker and other phases within callbacks.
Each phase once successful sends a message to itself to move onto the next phase,
allowing attempts to be retried.