Add kafka_server_connect callback to servers
This removes the connection from the init/1 callback,
so it can retry.

As this module is supervised by the main OTP app,
if it returns stop too many times it will stop the app.
eg {:stop, reason} can not be used.

Could consider :ignore, but returns `{:ok, ?}`
which means start_worker cannot pattern match